+++ src/window_manager.c
@@ -434,8 +434,8 @@ static void window_manager_create_window_proxy(int animation_connection, struct 
     SLSSetWindowOpacity(animation_connection, proxy->id, 0);
     SLSSetWindowResolution(animation_connection, proxy->id, 2.0f);
     SLSSetWindowAlpha(animation_connection, proxy->id, 1.0f);
-    SLSSetWindowLevel(animation_connection, proxy->id, g_layer_normal_window_level);
-    SLSSetWindowSubLevel(animation_connection, proxy->id, proxy->level);
+    SLSSetWindowLevel(animation_connection, proxy->id, proxy->level);
+    SLSSetWindowSubLevel(animation_connection, proxy->id, proxy->sublevel);
     proxy->context = SLWindowContextCreate(animation_connection, proxy->id, 0);
 
     CGRect frame = { {0, 0}, proxy->frame.size };
@@ -475,6 +475,7 @@ static void *window_manager_build_window_proxy_thread_proc(void *data)
     struct window_animation *animation = data;
 
     animation->proxy.level = window_level(animation->wid);
+    animation->proxy.sublevel = window_sublevel(animation->wid);
     SLSGetWindowBounds(g_connection, animation->wid, &animation->proxy.frame);
     animation->proxy.image = SLSHWCaptureWindowList(g_connection, &animation->wid, 1, (1 << 11) | (1 << 8));
     window_manager_create_window_proxy(g_connection, &animation->proxy);
@@ -571,6 +572,7 @@ void window_manager_animate_window_list_async(struct window_capture *window_list
             context->animation_list[i].proxy.frame.size.width  = (int)(existing_animation->proxy.tw);
             context->animation_list[i].proxy.frame.size.height = (int)(existing_animation->proxy.th);
             context->animation_list[i].proxy.level             = existing_animation->proxy.level;
+            context->animation_list[i].proxy.sublevel          = existing_animation->proxy.sublevel;
             context->animation_list[i].proxy.image             = CFRetain(existing_animation->proxy.image);
             __asm__ __volatile__ ("" ::: "memory");